Summation for iset.mm through sumss (#2846)
* Add ifiddc to iset.mm

This is ifid from set.mm with a decidability condition added.

* Add fser0const to iset.mm

This is an expression which comes up in at least one finite series
proof, and the theorem simplifies it.

* Add isumz to iset.mm

This is sumz from set.mm but with some changes in how we define a
summable set. The proof is similar to the set.mm one.

* Add cbvsum to iset.mm

The theorem is stated as in set.mm. The proof is very similar to the
set.mm one, it only needs the obvious changes for the differences
between df-sum and df-isum .

* Add cbvsumv to iset.mm

Copied without change from set.mm.

* Add cbvsumi to iset.mm

Copied without change from set.mm

* Add sumeq1i to iset.mm

Copied without change from set.mm.

* Add summation equality theorems to iset.mm

This is sumeq2i , sumeq12i , sumeq1d , sumeq2d , sumeq2dv ,
sumeq2ad , sumeq2sdv , 2sumeq2dv , sumeq12dv , and sumeq12rdv .
They are copied without change from set.mm

Remove the old sumeq2d from iset.mm. It corresponded to sumeq2dv
from set.mm but had a longer proof. It was unused in iset.mm so there is
no need to update any usages.

* Add sum2id and sumfc to mmil.html

* Remove znnen from mmil.html

This was proved in iset.mm some time ago. Minimize the iset.mm proof.

* Add sumfct to iset.mm

This is sumfc from set.mm with a different proof and a closure
condition.

* Add fmpttd , fmpt3d , fmptdf to iset.mm

Copied without change from set.mm.

* Add fsumf1o to iset.mm

The theorem is stated as in set.mm.  The proof is similar to the set.mm
one with the expected changes for difference in summation theorems and
other intuitionizing.

* Add ifandc to iset.mm

This is ifan from set.mm with a decidability condition added. The proof
is extremely similar to the set.mm one.

* Add isumss to iset.mm

This is like sumss from set.mm but needs a few additional hypotheses.
The proof is also similar to the set.mm one although it needs
intuitionizing on enough steps that some parts of it look moderately
different.