Add fisumcvg3 to iset.mm
This is fsumcvg3 from set.mm. The statement of the theorem adds a
decidability condition. The proof is fairly similar but parts of it are
revised to use an integer upper bound for the set A rather than a
supremum. The latter might be possible, given that we have the maximum
of two numbers and could, I suppose, generalize that to a finite set, but
the upper bound works fine for this proof.