Intuitionize summation from fsumsplit to sumpr
fsumsplit is stated as in set.mm. The proof needs intuitionizing for
decidability but is similar to the set.mm one.

Copy fsumsplitf from set.mm without modification.

sumsnf is stated as in set.mm. The proof is similar to the set.mm
one but needs intuitionizing for things like decidability. Fortunately,
the needed conditions are easy to satsify.

Add fsumsplitsn to mmil.html

Shorten proof of sumsn in set.mm. Also, copy it (without changes) to iset.mm.

Add fsum1 to iset.mm. Copied without change from set.mm.

sumpr is stated as in set.mm. The proof is similar to the set.mm one but
needs a small change related to finiteness of a pair.