Add fisumser to iset.mm
This is similar to fsumser in set.mm but needs fairly extensive changes
(along the lines of other finite sum theorems using seq ).

It requires that F be defined on ( ZZ>= ` M ) (in keeping with the
"partial sum of an infinite series" concept). The proof is similar with
changes relating to those changes and differences in seq and summation theorems.