Add three argument seq to iset.mm
The key insight here is that the three argument and four argument seq can
coexist (maybe for a transition period, or maybe longer if the four
argument version is needed early on in seq development) similarly to
the way that df-an and df-3an can both use the /\ constant.

This is the syntax axiom cseq3 and the df-seq3 definition.