brres and related theorems
Finishing the project which was initiated in PR 2831 along the logic of the conversation
https://github.com/metamath/set.mm/pull/2831#pullrequestreview-1109220740
with revision suggested by @benjub

Revision of brinxp is ready by the PR above (and now I filled in missing lines in changes-set.txt),
revision of brres is the topic of this PR.

Original brresALTV and opelresALTV was renamed to brres and opelres,
original brres and opelres was renamed to brresi and opelresi
(and original opelresi was renamed to opelidres).

Some (cca. 1/3 of the 100+) related theorems was minimized, at least with ancom ancoms ancomd biancom or similar.
We could shorten relexpindlem further with syl2an2r but I did not find its place.
In one case, dvres, I minimally changed, and shortened, its lemma, dvreslem.
In one case, setrec2lem2, I left a theorem in its proof, tz6.12-1, unchanged, which caused 1 step longer proof with ancoms.
It was new for me that minimize_with * /include_mathboxes does not find all options:
in case of relexpindlem, I shortened jca and mpd176 with mp2and, but this was not suggested by the program.