Add fsummulc2 to iset.mm
Stated as in set.mm.  The proof requires quite a bit of intuitionizing
even though it maintains the basic structure and many of the steps of
the original.  It is just that some of the theorems we depend on are
different and some parts some if expressions not present in the set.mm
proof.