Add modfsummod to iset.mm
Stated as in set.mm.  The proof is basically the same as the set.mm one
but needs changes including passing -. z e. A to the induction
step lemma ( modfsummods in set.mm, renamed to modfsummodlemstep here).
Putting the lemma in deduction form shortens a lot of proof steps, and
many steps need to have real changed to rational in modulus theorems
(which are just dealing with integers here anyway).

Although the lemma modfsummodslem1 is unchanged from set.mm, rename to
modfsummodlem1 to reflect the other naming here.