Update seq usage in resqrex in iset.mm
Remove the hypothesis from resqrexlemp1rp and resqrexlem1arp which
defines the sequence as it is unused there.

Change the definition of the sequence (in resqrex and the remaining
lemmas) to use df-seq3 syntax rather than df-iseq syntax.

Update resqrexlemf , resqrexlemf1 , and resqrexlemfp1 to use
theorems which refer to the df-seq3 syntax.

The remaining lemmas do not need to change as there is no change to the
compressed proof if they are just accepting the hypothesis and passing
it along to other lemmas without looking at its contents.