Add fidcenum to iset.mm
The forward direction - fidcenumlemim - is mostly already proved.

The reverse direction is fidcenumlemr .  This is adapted from the
AczelRathjen proof but because we have a variety of theorems regarding
finite sets to build on, we don't need to explicitly construct
bijections with natural numbers.

Includes lemmas fidcenumlemrk and fidcenumlemrks also based on
AczelRathjen .

Have the comment also cite the AczelRathjen definition of "finitely
enumerable".