Add exp3val to iset.mm
This is based on iexpval but uses the df-seq3 syntax rather than the
df-iseq syntax.  The proof is also a bit rearranged to be more
consistently in deduction form.  I didn't check whether that ended up
making it shorter, but it did make it easier to write and hopefully
makes it shorter or clearer as well.

Includes lemma exp3vallem which is similar to iexpvallem .