Prove expcnv in iset.mm
Stated as in set.mm.  The proof starts with expcnvap0 and removes the
restriction that A has to be apart from zero.

Includes expcnvre which is basically a lemma for expcnv but which I
suppose could conceivably be useful on its own.

Remove expcnv from missing theorems list in mmil.html