Add dfseq3-2 to iset.mm
This is df-iseq with S changed to _V .  Eventually this will replace
df-iseq and df-seq3 as the definition of seq but for now it is a
theorem proved from them.

Update iset-discouraged because df-iseq is marked as discouraged and
the dfseq3-2 proof adds one usage.  This is OK because the purpose of
dfseq3-2 is to work towards removing df-iseq .