"Minus times minus is plus"
recently, I stumbled over the citation "Minus times minus is plus, The reason for this we need not discuss." in the Metamath book (p. 5). I wondered if this statement is proved in set.mm, and found that it isn't. Therefore I formalized it and wproved it, see ~nnmtmip and ~rpmtmip.
Having these theorem, this statement actually "needs not be discussed" anymore. In the comments of these two theorems, I also explained "The reason for this".

Although the proofs of the two theorems are short and straightforward, I think it is more than appropriate to have them explicitly in set.mm, just to avoid the arguments "that's trivial" or even "that easily follows from ~mul2neg" which are criticised in the Metamath book (and by Norm himself, which was one of the starting points to "invent" Metamath). As a consequence, each theorem/matemathical statement  made in the Metamath book should be proved in set.mm.