Add apdivmuld to iset.mm
This is an analogue of divmulapd ( divmuld in set.mm) but for apartness
rather than equality.  In general, where there is a theorem of the form
X = Y <-> Z = W we often want X =//= Y <-> Z =//= W so make a note of
that in the section on intuitioniztion in mmil.html.