Add tanaddap to iset.mm
This is tanadd from set.mm with not equal changed to apart.  The proof
needs a lot of intuitionizing to change not equal to apart, or change
equal to apart (since we can't convert equal to apart at the end, the
way the set.mm proof converts equal to not equal).

Includes lemma tanaddaplem