Document theorems implementing CPCTC.
Many geometry textbooks refer to the theorem congruent triangles
are congruent (CPCTC).  This commit adds notes to various comments
to specifically indicate the various theorems that together constitute
CPCTC.

In some sense this completes #49, which identifies CPCTC as an
example of what it means to add "support for Euclidean Geometry".

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>