Add fiintim to iset.mm
This is based on fiint from set.mm but is only one direction of the
biconditional.  The proof is based on the set.mm proof of fiint but
needs some intuitionizing to replace case elimination with the theorem
(which we have) that a natural number is either zero or greater than
zero.