Add strnfv2d to iset.mm
This is strfv2d from set.mm with a ( E ` ndx ) e. NN condition added.
The proof needs some intuitionizing but is basically the set.mm proof.

Minimize strfv2d in set.mm and adjust comment to refer to strfv2 rather
than strfv (because this theorem is more like strfv2 in terms of how its
hypotheses are written).