Add strsetsid to iset.mm
This is based on setsidvald from set.mm, but various hypotheses are
tweaked (most notably that S is a structure) and various parts of the
proof are intuitionized.

Replaces setsssvald which didn't reflect how we expect to use this (for
example, there is no need to cater to slot numbers which are not
integers).