Update algorithm theorems in iset.mm
This is mostly to use df-seq3 syntax.

ialgrlem1st and ialgrlemconst do not need to change but adjust their
comments to refer to the current seq theorems.

Modify ialgr0 to use the df-seq3 syntax and also remove the now-unnecessary
S e. V hypothesis.

Copy algrf from set.mm without changes which is now possible because we
have theorems such as seqf .  Replaces ialgrf .

algrp1 is stated as in set.mm but the proof needs modifications
(relative to the set.mm proof) to use seq3p1 instead of seqp1 and to
replace fvex with various replacements.  Replaces ialgrp1 .

Copy alginv from set.mm without change, which is now feasible because we
now have algrf and algrp1 .  Replaces ialginv .

Add algcvg which is stated as in set.mm.  The proof is
the same as the set.mm proof with the very minor change to use algr3-0 .
Replaces ialgcvg .

Copy algcvga and algfx from set.mm without change.  Replaces ialgcvga and ialgfx .

Copy eucalgcvga from set.mm without modification
(replaces eucialgcvga ).

Copy eucalg from set.mm - the only change needed is to use algr3-0
instead of algr0 (replaces eucialg ).  Update mm_100.html with the
changed name of eucalg .

Remove entries from mmil.html for theorems which now are stated the same
in set.mm and iset.mm.

Remove entries in iset-discouraged for discouraged theorems which are no
longer used because of these changes.