Reword comments in iset.mm concerning countable
In set.mm, it is OK for dominated by _om and countable to be synonymous.
But in iset.mm the usual definition of countable has to do with whether
there is a map from _om onto the set (more precisely, the disjoint
union of 1o and the set, to make the empty set countable).