Add xaddval to iset.mm
Stated as in set.mm.  The proof needs intuitionizing around decidable
propositions and ovex but is otherwise similar to the set.mm proof.

Along those lines, add xrpnfdc and xrmnfdc which are used in this proof
and will likely also be needed in a lot of the extended real theorems.

Includes xaddf which is stated as in set.mm.  The proof is somewhat
similar to the set.mm proof but needs intuitionizing in quite a few
steps.