Add negated equality theorems to mmil.html
These are theorems which are doing forms of contraposition or case
elimination which only hold for decidable propositions.  They are
from necon1ad to pm2.61da3ne .  They had been listed in git logs and are
now being listed in mmil.html too.

In iset.mm, rename necon2bbii to necon2bbiidc . The latter had always
been intended and matches the convention of having different names if
the theorem in set.mm is different.