Add xrmaxiflemlub to iset.mm
This is the extended real analogue to maxabslemlub .

The proof involves dividing into cases based on elxr and similar
theorems but is not difficult.

Includes lemma xrmaxiflemcl for closure of the expression which we will
prove to be the supremum.

Includes lemma xrmaxiflemval which corresponds to maxabslemval for reals
and which basically says that this expression is a supremum.

Includes lemma xrmaxiflemcom which shows that this expression is
commutative.