Add setsmsbasg to iset.mm
This is setsmsbas from set.mm with additional set existence hypotheses.
The proof is similar to the set.mm proof.

Update iset-discouraged for the additional usage of basendx .  Because
we are using it to show that ( Base ` ndx ) =/= ( TopSet ` ndx ) we need
to look at the index one way or another.