Add bdbl to iset.mm
This is stdbdbl from set.mm with minimum expressed as the infimum of a
pair.  The proof needs intuitionizing not just for the minimum theorems
but also to revamp the use of < and <_ to reflect what is possible
there.  Fortunately, the overall logic of the proof works with those
changes.