Add bdxmet to iset.mm (#3194)
* Add bdtri to iset.mm

This is a triangle inequality stated in terms of minimum.

Includes lemma bdtrilem which contains many of the calculations.

* Add xrminrecl to iset.mm

This is similar to xrmaxrecl and is readily proved in terms of xrmaxrecl

* Add xrbdtri to iset.mm

This is the extended real version of bdtri and is proved by the usual
decomposition into cases using elxr and similar theorems.

* Add bdxmet to iset.mm

This is based on stdbdxmet from set.mm.  Although the proof more or less
follows the structure of the set.mm proof, there is some extensive
intuitionizing to do in places.