Change sum_ definition in iset.mm
This gets rid of df-isum (which uses the df-iseq syntax for seq) and
replaces it with df-sumdc (which is the same but uses df-seq3 syntax).

Revise the proof of sumeq1 , nfsum1 , nfsum , sumeq2 , and cbvsum
to be in terms of the new definition.

For now, keep dfsumdc but prove it in terms of df-sumdc .  Since it is
exactly the same as df-sumdc with a now-unneeded antecedent, we'll be
able to get rid of it soon but keeping it for the moment will prevent us
from having to change everything at once.

Update mmil.html to refer to df-sumdc where it had been referring to
df-isum .