Intuitionize section "Continuity in metric spaces" (#3203)
* Add metcnp3 to iset.mm

Copied without change from set.mm.

* Add metcnp to iset.mm

Copied without change from set.mm

* Add metcnp2 and metcn to iset.mm

Copied without change from set.mm.

* Add metcnpi to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Remove epsilon-delta comments in set.mm

The comment was "TODO - delete this if not used?" and was on metcnpi2
and metcnpi3 .  metcnpi2 is used in ftc1lem6 and ftc1cnnc . metcnpi3
is used in blocnilem .

* Add metcnpi2 to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add metcnpi3 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.