Intuitionize section "Topology on the reals" through metdsf (#3208)
* Add ad4ant123 etc to iset.mm

This is ad4ant123 , ad4ant124 , ad4ant134 , and ad4ant234 .
Copied without change from set.mm.

* Add iooinsup to iset.mm

This is iooin from set.mm but expressing maximum as the supremum of a
pair and minimum as infimum of a pair.  The proof is taken from the
set.mm proofs of iooin and ixxin and follows that proof fairly closely.

* add dscmet , dscopn to mmil.html

* Add qtopbasss to iset.mm

This is qtopbaslem from set.mm with additional hypotheses saying that S
is closed under maximum and minimum (which is trivial in set.mm but
which is not so simple for us).

* Add minclpr to iset.mm

This is maxclpr but for minimum and is readily proved in terms of
maxclpr .

* Add qtopbas to iset.mm

Stated as in set.mm.  The proof invokes qtopbasss (as the set.mm proof
invokes qtopbaslem ) but needs to pass two additional hypotheses.

* Add retopbas to iset.mm

Stated as in set.mm.  The proof is basically the set.mm one with some
intuitionizing for the differences between qtopbaslem and qtopbasss .

* Add retop and uniretop to iset.mm

Copied without change from set.mm.

* Add retopon and retps to iset.mm

Copied without change from set.mm.

* Add iooretopg to iset.mm

This is iooretop from set.mm except that A and B need to be extended
reals.  The proof is also the set.mm one with that being the only needed
change.

* Add icccld , icopnfcld , iocmnfcld to mmil.html

* add qdensere to mmil.html

* Add cnmetdval to iset.mm

Copied without change from set.mm.

* Add cnmet and cnxmet to iset.mm

Copied without change from set.mm.

* Add cnbl0 and cnbl0 to iset.mm

Copied without change from set.mm.

* Add remetdval to iset.mm

Copied without change from set.mm.

* Add remet and rexmet to iset.mm

Copied without change from set.mm.

* Add bl2ioo and ioo2bl to iset.mm

Copied without change from set.mm.

* Add ioo2blex to iset.mm

Copied without change from set.mm.

* Add blssioo to iset.mm

Stated as in set.mm.  The proof needs intuitionizing at a number of
steps but is basically the set.mm proof.

* Add tgioo to iset.mm

Stated as in set.mm.  The proof requires intuitionizing in various
places but is basically the set.mm proof.

* add qdensere2 to mmil.html

* add blcvx to mmil.html

* Add tgqioo to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add resabs1d to iset.mm

Copied without change from set.mm

* Add resubmet to iset.mm

Copied without change from set.mm

* Add df-cnfld and CCfld to mmil.html

* Add df-ordt , df-xrs to mmil.html

* Add zcld to mmil.html

* Add iccntr to mmil.html

* add df-conn to mmil.html

* Add opnreen to mmil.html

* Add rectbntr0 to mmil.html

* Add cnmpt1ds to mmil.html

* add xrsupss to mmil.html

* add supxrcl to mmil.html

* Add infxrcl , metdsval , metdsf to mmil.html