Add expcncf to iset.mm
Stated as in set.mm.  The proof is by induction and is a little bit like
the set.mm proof of expcn but the details are different.

Add mulcncf to iset.mm.  Stated as in set.mm.  The proof is a new one
based on mulcn2 .

Includes lemma mulcncflem which contains part of the proof of mulcncf .