Intuitionize continuous function theorems from cncfmet to expcncf (#3218)
* Add cncfmet to iset.mm

Copied without change from set.mm

* Add cncfmptc to iset.mm

Stated as in set.mm.  The proof is a new one based on elcncf1di .

* Add cncfmptid to iset.mm

Stated as in set.mm.  The proof is a new one using elcncf1di .

* Add cncfmpt1f to iset.mm

Copied without change from set.mm

* Add addccncf to iset.mm

Stated as in set.mm.  The proof is a new one based on elcncf1di .

* Add reccn2ap to iset.mm

This is reccn2 from set.mm with not equal changed to apart and minimum
expressed as the infimum of a pair.

* Add cdivcncfap to iset.mm

This is cdivcncf from set.mm with not equal changed to apart.  The proof
is a new one based on cncfco , mulc1cncf , and reccn2ap .

* Add negcncf to iset.mm

Stated as in set.mm.  The proof is a new one based on elcncf1di .

* Add negfcncf to iset.mm

Copied without change from set.mm.

* Add expcncf to iset.mm

Stated as in set.mm.  The proof is by induction and is a little bit like
the set.mm proof of expcn but the details are different.

Add mulcncf to iset.mm.  Stated as in set.mm.  The proof is a new one
based on mulcn2 .

Includes lemma mulcncflem which contains part of the proof of mulcncf .