misc shortenings, mteqand, resubdi, fltne
eldifsnneq was originally used in prjspreln0 but then I used mteqand

I shortened eldifsnneq using eldifsn but then the auto minimizer found 
eldifsni and the proof is obliterated. The minimizer doesn't find 
eldifsni by itself though.

fiminre was originally changed to reduce axioms but I put proof 
shortened because the proof is really just the parallel of fimaxre

fltne was made because I thought I found a simple proof of FLT but of 
course it didn't work out.