Swap idi and a1ii within both set.mm and iset.mm
Theorem `idi` is conceptually more primitive than `a1ii`,
so it should be first.  Theorem `idi` is literally just a copy,
while `a1ii` is a selection of a copy.

It doesn't *matter* from the point-of-view of truth, but
when you're explaining each assertion in sequence, it's easier
to explain this pair of theorems if `idi` comes first.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>