Add seqf2 to iset.mm (#3287)
* Add seqvalcd to iset.mm

Similar to seq3val but with two types for the terms and the accumulating
value.

* Add seqf2 to iset.mm

Stated as in set.mm.  The proof is based on the iset.mm proof of seqf .

In mmil.html, describe seqf2 as the replacement for seqcl2 as that is
presumably as close as we can get at least given the current state of
seq and the theorems it depends on.