iset.mm: Shroeder-Bernstein fares little better if one set is _om (#3292)
* Add xp01disjl to set.mm and iset.mm

This is similar to xp01disj but with the singletons on the left hand
side.  This is what you'll want in dealing with disjoint unions where we
put the singletons on the left.

* Add uncdadom and pm110.643 to mmil.html

* Add djudoml to iset.mm

This is A ~<_ ( A |_| B ) and is readily proved from existing inl theorems.

* Add djudomr to iset.mm

This is B ~<_ ( A |_| B ) and is an easy consequence of existing inr
theorems.

* Add caseinr to iset.mm

Just like caseinl but for the right injection.

The proof is very similar to the caseinl proof.

* Move nnsucpred from mathbox to main in iset.mm

* Add inl11 to iset.mm

This says that inl is one-to-one.  The proof is easy from the definition
and opth .

* Add omp1eom to iset.mm

Adding one to _om produces a result equinumerous to _om .

Includes lemma omp1eomlem

* Add endjusym to iset.mm

Reversing right and left operands of a disjoint union.  The proof is
relatively straightforward.  This proof requires the operands to be
sets.

* Add sbthom to iset.mm

The proof was written informally in a github comment by Mario Carneiro
and I formalized it.  It shows that we cannot prove the Schroeder-Bernstein
theorem even for the case where one set is _om .

Includes lemma sbthomlem