Rename df-mpt2 to df-mpo
This is in both set.mm and iset.mm and is accompanied by a plan to
rename all the other theorems which contain this abbreviation.

There has been a lot of discussion especially in github of how to name
this but in a nutshell the reason for the change is that "2" does not
convey anything meaningful, but "o" stands for "operation" and thus
makes sense here.