Add difinfinf to iset.mm
This is based on difinf from set.mm.  The first change is how we express
that a set is infinite.  The second is that we seem to need decidable
equality.  It seems counterintuitive to need it but I'm not sure
it can be removed.  The third change is that the finite set being
subtracted must be a subset of the set it is being subtracted from,
which is pretty common in set subtraction theorems in iset.mm.

Includes difinfsnlem and difinfsn which are for the case where the
finite set being subtracted is a singleton (the hard part - after that
difinfinf is just an easy finite set induction).