Fix metamath-knife caching
The symptom was a "Path(s) specified in the action for caching do(es) not
exist, hence no cache is being saved" error message.

There would seem to have been a few problems in how it was set up.  Will
have to see whether this is sufficient to get it to work.

Also update location of metamath-knife repository