Don't use "regular union" in comments
When talking about union versus disjoint union it is tempting to use
"regular union", "plain union", or some other qualifier for the former,
but it probably is clearer to just say "union" as we would when we
aren't contrasting with disjoint union.  The sentence structure
hopefully makes the contrast clear.