Fixes to metamath-knife job in the automated checks (#3407)
* Also verify markup with metamath-knife

The rules are slightly different than metamath-exe but they are good
ones.  Fix three small things in iset.mm so these checks will pass.

Also add --parse-typesetting which checks at least some aspects of the
$t syntax.

* Fix metamath-knife caching

The symptom was a "Path(s) specified in the action for caching do(es) not
exist, hence no cache is being saved" error message.

There would seem to have been a few problems in how it was set up.  Will
have to see whether this is sufficient to get it to work.

Also update location of metamath-knife repository