Start the process of moving set.mm from +c to |_| (#3406)
* Expand df-xps entry in mmil.html

Have the text say more about what it depends on.

* Add df-cda , cdafn , and cdaval to mmil.html

In this case disjoint union in iset.mm has equivalents.

* Add undjudom to set.mm

This is uncdadom but using the disjoint union syntax.

* Say that df-dju is preferred over df-cda

This is in the comment for cdaval where we discuss our approach to
cardinal arithmetic.  There are no advantages to df-cda over df-dju so
let's begin the process of converting to df-dju .

* Add endjudisj to set.mm

The theorem is stated as in iset.mm.  The proof is based on the set.mm
proof of cdaun .

Move the iset.mm proof of endjudisj so that it is in a similar place in
the file to the set.mm version.

* Add djuen to set.mm

Stated as in iset.mm.  The proof is based on the set.mm proof of cdaen .

Move the iset.mm djuen to be in a similar place in the file as the
set.mm one.

* Add djuenun

This is cdaenun with +c changed to |_| .  The proof is based on the
set.mm proof of cdaenun .

Copy without change to iset.mm.

* Don't use "regular union" in comments

When talking about union versus disjoint union it is tempting to use
"regular union", "plain union", or some other qualifier for the former,
but it probably is clearer to just say "union" as we would when we
aren't contrasting with disjoint union.  The sentence structure
hopefully makes the contrast clear.

* Add cdadju to set.mm

This is a transition aid as we move from +c to |_|