Add djudom1 to set.mm
This is similar to cdadom1 but for disjoint union.  It adds a sethood
hypothesis for C (due to differences in proper class handling between
df-cda and df-dju ).  Other than that, the proof is similar to the proof
of cdadom1 .

Mark cdadom1 as discouraged.  Note cdadom1 and djudom1 in mmil.html