Split metamathexe and tex jobs
The metamath.exe job has acquired a lot of semi-related functionality
which is OK up to a point but the script is getting a bit hard to read
and perhaps more importantly, the tex job (in particular, installing the
tex packages) is a lot slower than the other checks.  So by splitting it
out we can speed up the whole run a bit, and get faster feedback if
there is an error.