Provide |_| versions of +c theorems from cdadom3 to pwcdaidm (#3468)
* Add djudoml to set.mm

Stated as in iset.mm but with a different proof based on the set.mm
proof of cdadom3 .  Mark cdadom3 as discouraged (it it the same thing
with +c rather than |_| ).

* Add djuxpdom to set.mm

Similar to cdaxpdom with +c changed to |_| .  Mark cdaxpdom as
discouraged and add cdaxpdom and djuxpdom to mmil.html.

* Add djufi to set.mm

Like cdafi but using |_| notation rather than +c .

Add cdafi and djufi to mmil.html.  Mark cdafi as discouraged.

* Add djuexb to set.mm and iset.mm

The proof is similar, but not identical, between set.mm and iset.mm.

Also mention xpexcnv in mmil.html.

* Add djuinf to set.mm

This is cdainf with +c changed to |_| .  The proof is similar to the
proof of cdainf although some changes are needed to deal with
differences in handling of proper classes.

* Add infdju1 to set.mm

This is infcda1 with +c changed to |_| .  The proof needs few changes
from the infcda1 proof.

Mark infcda1 as discouraged.

Add infdifsn , infcda1 , infdju1 to mmil.html

* Add pwdju1 to set.mm

This is pwcda1 with +c changed to |_| .  The proof is very similar.

Mark pwcda1 as discouraged.

Add pwcda1 and pwdju1 to mmil.html

* Expand comments on pwpw0 in iset.mm

This is the comment of pw1dom2 in iset.mm and the entry for pwpw0 in
mmil.html.

* Add pwdjuidm to set.mm

This is pwcdaidm using |_| instead of +c .  The proof is similar.

Mark pwcdaidm as discouraged.

Add to mmil.html.