Expand mathbox guidelines
The new text in set.mm is particularly around abandoned mathboxes, but
the idea of moving important theorems to main (eventually) is a bit
more general than that.

In the case of iset.mm, I think what had been there was just an old copy
of the set.mm mathbox guidelines.  It seems better to just have a few
introductory sentences and refer people to the set.mm guidelines,
rather than try to keep two copies in sync.