Fix df-tsk comment.
Because underscore is followed by ampersand, metamath-exe is not
treating this as italics and latest metamath-knife is giving an error.
Apparently we need to either give up the italics or give up the umlaut,
and turning the umlaut into a following e is a pretty standard thing to
do, so I guess we do that.