Add df-iress to iset.mm
This is similar to df-ress from set.mm but only includes the
non-trivial case.  Given reasonable conditions (for example
those in strsetsid ) the two cases give the same results, and
df-ress is infeasible in iset.mm because it wouldn't typically
be easy to show that ( Base ` w ) C_ x is decidable.

Remove df-ress from iset.mm; as described above it would be hard to work
with and would not have any of the advantages it has in set.mm.

- Add ressvalsets to iset.mm

This is similar to ressval in set.mm but adjusted for the difference
between set.mm's df-ress and iset.mm's df-iress .

- Add ressex to iset.mm

This replaces ovex but for |`s

- Add ressval2 to iset.mm

Stated as in set.mm but the proof in iset.mm is trivial from ressvalsets .
(In fact, we retain ressval2 only for consistency with set.mm).

- Add ressval3d to iset.mm

Stated as in set.mm.  The proof is pretty simple because of df-iress
(and in fact does not need the Fun S and E e. dom S hypotheses, but they
are kept for consistency with set.mm).

- Add ressid2 to mmil.html.  Remove from iset.mm.

- Add strressid to iset.mm

This is ressid from set.mm with additional conditions on the structure.
Replaces the ressid which had been in iset.mm which relied on the
previous definition of structure restriction.