Finish removing +c from set.mm (#3558)
* Revise dmdprdpr and dprdpr in set.mm

Change from +c to a pair of ordered pairs in stating the theorems and in
the proofs.  Other than that the proofs are unchanged.

* Remove xpscfcda from set.mm

No longer used and replaced by xpscf

* Remove xpsfrnel2cda from set.mm

No longer used and replaced by xpsfrnel2

* Remove xpsc1 from set.mm

No longer used and replaced by fvpr1o

* Remove xpsc0 from set.mm

No longer used and replaced by fvpr0o

* Remove xpscfn from set.mm

No longer used and replaced by fnpr2o

* Revise df-xps in set.mm

Change the definition to use a pair of ordered pairs instead of +c .
This is the same as what had been a theorem dfxpspr so remove that
theorem and update its one usage to refer to df-xps .

* Remove xpscg from set.mm

No longer used.

* Remove xpsc from set.mm

No longer used.

* Remove cdadju from set.mm

As there is no +c left we can just recommend |_| .

* Remove cdaval from set.mm

Move most of its comment to a section header for the Cardinal number
arithmetic section (with some tweaks to the wording).

cdaval has no more usages so we can just refer people to df-dju instead.

* Remove df-cda and +c from set.mm

No longer used.  Remove the definition and the syntax and recommend
df-dju instead.