Add txmetcnp to iset.mm
Stated as in set.mm.  The proof is a new one, because the set.mm proof
is in terms of a bunch of notations which we don't have (at least not yet).
The proof is vaguely similar to the set.mm proof in terms of being
expressed in terms of the maximum metric.