df-selv theorems (#3615)
* move and fix df-selv

df-selv applies to any polynomial. It is basically a way to further show how evalSub works, so it is moved up

df-selv intended the definition of algSc but subtly messed up, which is fixed

* move csbeq* theorems

simplify df-selv using the variable s

* csbeq12dv; the value of df-selv

fix df-selv even more)
(explanation of df-selv change:
evalSub takes a subring. Since a subring is a subset of the base ring,
we use the base set instead of the df-imas-structure)

df-selv may be changed further in future commits, since I don't understand it

* fix df-selv more, df-selv lemmas

Fix df-selv again, simply another missing Base call

The lemmas (currently in my mathbox) show that df-selv at least
provides arguments in the domain of df-evls

Move cslv up, to be consistent with where df-selv now is

Several theorems moved or made in order to prove the lemmas

(the proof of imasasclsubrglem is very similar to prjspertr)