Add dvaddxxbr to iset.mm
This is dvaddbr from set.mm with a change to the syntax of the topology
on the complex numbers and in which X and Y are replaced by X .  This avoids
a lot of steps related to ( X i^i Y ) some of which rely on excluded
middle.  The proof is similar to the set.mm proof but various parts are
intuitionized or simplified.