Add product rule for derivatives to iset.mm (#3674)
* Add dvmulxxbr to iset.mm

This is dvmulbr from set.mm with a few changes to the hypotheses.  The
proof is similar to the set.mm proof but intuitionized in various
places.

* Add dvmulxx to iset.mm

This is dvmul from set.mm with slight changes to the hypotheses.

* Add dvimulf to iset.mm

This is dvmulf with a small change to the hypotheses.  The proof is
similar to the set.mm proof but does need some intuitionization.